refactor: reuse code for edit and new reservation
