+++ apps/admin-ui/src/hooks/index.ts
@@ -2,8 +2,6 @@ export { useReservationUnitTypes } from "./useReservationUnitTypes";
 export { useReservationUnitOptions } from "./useReservationUnitOptions";
 export { useUnitOptions } from "./useUnitOptions";
 export { useOptions } from "./useOptions";
-export { useApplicationFields } from "./useApplicationFields";
-export { useGeneralFields } from "./useGeneralFields";
 export { useCheckPermission } from "./useCheckPermission";
 export { useCheckCollisions } from "./useCheckCollisions";
 export { useDenyReasonOptions } from "./useDenyReasonOptions";
